parseq_requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
ax-platform==0.2.5.1
botorch==0.6.4
cachetools==5.2.0
#certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
click==8.0.4
cycler==0.11.0
distlib==0.3.6
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
fvcore==0.1.5.post20221122
google-auth==2.14.1
google-auth-oauthlib==0.4.6
gpytorch==1.8.1
grpcio==1.43.0
huggingface-hub==0.11.0
hydra-core==1.2.0
idna==3.4
imageio==2.22.4
imgaug==0.4.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
iopath==0.1.10
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
kiwisolver==1.4.4
lmdb==1.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
msgpack==1.0.4
multidict==6.0.2
multipledispatch==0.6.0
networkx==2.6.3
nltk==3.7
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.2.3
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
plotly==5.11.0
portalocker==2.6.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyparsing==3.0.9
pyro-api==0.1.2
pyro-ppl==1.8.0
pyrsistent==0.19.2
python-dateutil==2.8.2
pytorch-lightning==1.6.5
pytz==2022.6
PyWavelets==1.3.0
PyYAML==6.0
ray==1.13.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Shapely==1.8.5.post1
six==1.16.0
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==2.1.1
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.6.11
torchmetrics
tqdm==4.64.1
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.16.7
Werkzeug==2.2.2
yacs==0.1.8
yarl==1.8.1
zipp==3.10.0
